+++ lib/test/drm/playready/utils/stringToUtf16Buffer.ts
@@ -5,4 +5,6 @@ import { stringToUint16 } from '@svta/common-media-library/utils/stringToUint16.
  * @param input - The input string to encode.
  * @returns The UTF-16 encoded ArrayBuffer.
  */
-export const stringToUtf16Buffer = (input: string): ArrayBuffer => stringToUint16(input).buffer as ArrayBuffer;
+export function stringToUtf16Buffer(input: string): ArrayBuffer {
+	return stringToUint16(input).buffer as ArrayBuffer;
+}