+++ lib/test/iso/bmff/moov.test.ts
@@ -1,8 +1,8 @@
-import { assert, describe, it, parseContainer } from './util/box.ts';
+import { assert, describe, it, parseContainer, type MovieBox } from './util/box.ts';
 
 describe('moov box', function () {
 	it('should correctly parse the box', function () {
-		const box = parseContainer('captions.mp4', -3);
+		const box = parseContainer('captions.mp4', -3) as MovieBox;
 
 		assert.ok(box);
 		assert.strictEqual(box.type, 'moov');