+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Changed
+
+- Update Node.js development version to 24
+  ([#211](https://github.com/streaming-video-technology-alliance/common-media-library/issues/211))
+
 ## [0.13.0] - 2025-06-09
 
 ### Added