+++ CHANGELOG.md
@@ -12,12 +12,14 @@ and this project adheres to
 
 - Add map type for ISO BMFF boxes to their type properties
   ([#208](https://github.com/streaming-video-technology-alliance/common-media-library/issues/208))
+- Added Zlema Throughput Estimator
+  ([#212](https://github.com/streaming-video-technology-alliance/common-media-library/pull/212))
 
 ### Changed
 
 - Update Node.js development version to 24
   ([#211](https://github.com/streaming-video-technology-alliance/common-media-library/issues/211))
-- Added Zlema Throughput Estimator ([#212](https://github.com/streaming-video-technology-alliance/common-media-library/pull/212))
+
 
 ## [0.13.0] - 2025-06-09
 
@@ -29,8 +31,10 @@ and this project adheres to
   ([#160](https://github.com/streaming-video-technology-alliance/common-media-library/pull/160))
 - VTT Parsing Utilities
   ([#193](https://github.com/streaming-video-technology-alliance/common-media-library/issues/193))
-- Added Harmonic Mean Throughput Estimator ([#199](https://github.com/streaming-video-technology-alliance/common-media-library/pull/199))
-- Added Arithmetic Mean Throughput Estimator ([#201](https://github.com/streaming-video-technology-alliance/common-media-library/pull/201))
+- Added Harmonic Mean Throughput Estimator
+  ([#199](https://github.com/streaming-video-technology-alliance/common-media-library/pull/199))
+- Added Arithmetic Mean Throughput Estimator
+  ([#201](https://github.com/streaming-video-technology-alliance/common-media-library/pull/201))
 
 ### Fixed