+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to
 
 - Update Node.js development version to 24
   ([#211](https://github.com/streaming-video-technology-alliance/common-media-library/issues/211))
+- Fixed TypeScript errors related to Boxes ([#214](https://github.com/streaming-video-technology-alliance/common-media-library/pull/214))
 
 ## [0.13.0] - 2025-06-09